ForOffice/image_to_excel_converter/requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.16.4
openpyxl==2.6.2
Pillow==6.0.0
PySide2==5.13.0
shiboken2==5.13.0
